requirements.txt
streamlit==1.28.1
pypdf2==3.0.1
python-docx==0.8.11
openai==1.3.7
google-generativeai==0.3.2
neo4j==5.14.1
python-dotenv==1.0.0
pydantic==2.5.0
typing-extensions==4.8.0
requests==2.31.0
nltk==3.8.1
pandas==2.1.4
beautifulsoup4==4.12.2   
sentence-transformers==2.2.2
huggingface-hub==0.24.6
scikit-learn==1.3.2
